requirements.txt
langchain
langgraph
langchain_openai
langchain_community
pymongo
pinecone-client
sentence-transformers
streamlit
python-dotenv
fastapi
uvicorn
motor 